Moved tsc time for tdx code to openhcl_boot.
