+++ openhcl/minimal_rt/src/arch/aarch64/reftime.rs
@@ -2,12 +2,9 @@
 // Licensed under the MIT License.
 
 use super::hypercall::get_register_fast;
-use crate::isolation::IsolationType;
 
-pub fn reference_time(_isolation: IsolationType) -> Option<u64> {
-    Some(
-        get_register_fast(hvdef::HvArm64RegisterName::TimeRefCount.into())
-            .expect("failed to query reference time")
-            .as_u64(),
-    )
+pub fn reference_time() -> u64 {
+    get_register_fast(hvdef::HvArm64RegisterName::TimeRefCount.into())
+        .expect("failed to query reference time")
+        .as_u64()
 }
