+++ openhcl/minimal_rt/src/arch/x86_64/mod.rs
@@ -11,4 +11,3 @@ pub mod intrinsics;
 pub mod msr;
 pub mod reftime;
 pub mod serial;
-pub mod tdcall;
