+++ openhcl/minimal_rt/src/reftime.rs
@@ -2,9 +2,8 @@
 // Licensed under the MIT License.
 
 //! Hypervisor reference time.
-use crate::isolation::IsolationType;
 
 /// Returns the current reference time from the hypervisor, in 100ns units.
-pub fn reference_time(isolation: IsolationType) -> Option<u64> {
-    crate::arch::reference_time(isolation)
+pub fn reference_time() -> u64 {
+    crate::arch::reference_time()
 }
