+++ openhcl/openhcl_boot/src/arch/x86_64/mod.rs
@@ -13,9 +13,9 @@ pub mod tdx;
 mod vp;
 mod vsm;
 
+use crate::host_params::shim_params::IsolationType;
 pub use memory::setup_vtl2_memory;
 pub use memory::verify_imported_regions_hash;
-use minimal_rt::isolation::IsolationType;
 use safe_intrinsics::cpuid;
 pub use vp::setup_vtl2_vp;
 pub use vsm::get_isolation_type;
