+++ openhcl/openhcl_boot/src/dt.rs
@@ -4,6 +4,7 @@
 //! Module used to write the device tree used by the OpenHCL kernel and
 //! usermode.
 
+use crate::host_params::shim_params::IsolationType;
 use crate::host_params::PartitionInfo;
 use crate::host_params::COMMAND_LINE_SIZE;
 use crate::sidecar::SidecarConfig;
@@ -25,7 +26,6 @@ use loader_defs::shim::MemoryVtlType;
 use memory_range::walk_ranges;
 use memory_range::MemoryRange;
 use memory_range::RangeWalkResult;
-use minimal_rt::isolation::IsolationType;
 
 /// AArch64 defines
 mod aarch64 {
