+++ openhcl/openhcl_boot/src/host_params/dt.rs
@@ -6,6 +6,7 @@
 use super::shim_params::ShimParams;
 use super::PartitionInfo;
 use crate::boot_logger::log;
+use crate::host_params::shim_params::IsolationType;
 use crate::host_params::COMMAND_LINE_SIZE;
 use crate::host_params::MAX_CPU_COUNT;
 use crate::host_params::MAX_ENTROPY_SIZE;
@@ -24,7 +25,6 @@ use loader_defs::paravisor::CommandLinePolicy;
 use memory_range::subtract_ranges;
 use memory_range::walk_ranges;
 use memory_range::MemoryRange;
-use minimal_rt::isolation::IsolationType;
 
 /// Errors when reading the host device tree.
 #[derive(Debug)]
